Extract cli options to cli module
