+++ bin/full-node/cli.rs
@@ -1,0 +1,100 @@
+// Substrate-lite
+// Copyright (C) 2019-2020  Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+//! Provides the [`CliOptions`] struct that contains all the CLI options that can be passed to the
+//! binary.
+//!
+//! See the documentation of the [`structopt`] crate in order to learn more.
+//!
+//! # Example
+//!
+//! ```no_run
+//! use structopt::StructOpt as _;
+//! let cli_options = full_node::CliOptions::from_args();
+//! println!("Quiet: {:?}", cli_options.quiet);
+//! ```
+//!
+// TODO: I believe this example isn't tested ^ which kills the point of having it
+
+use std::path::PathBuf;
+
+// Note: the doc-comments applied to this struct and its field are visible when the binary is
+// started with `--help`.
+
+#[derive(Debug, structopt::StructOpt)]
+pub struct CliOptions {
+    /// Chain to connect to ("polkadot", "kusama", "westend", or a file path).
+    #[structopt(long, default_value = "polkadot")]
+    pub chain: CliChain,
+    /// No output printed to stderr.
+    #[structopt(short, long)]
+    pub quiet: bool,
+    /// Coloring: auto, always, never
+    #[structopt(long, default_value = "auto")]
+    pub color: ColorChoice,
+}
+
+#[derive(Debug)]
+pub enum CliChain {
+    Polkadot,
+    Kusama,
+    Westend,
+    Custom(PathBuf),
+}
+
+impl core::str::FromStr for CliChain {
+    type Err = core::convert::Infallible;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if s == "polkadot" {
+            Ok(CliChain::Polkadot)
+        } else if s == "kusama" {
+            Ok(CliChain::Kusama)
+        } else if s == "westend" {
+            Ok(CliChain::Westend)
+        } else {
+            Ok(CliChain::Custom(s.parse()?))
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum ColorChoice {
+    Always,
+    Auto,
+    Never,
+}
+
+impl core::str::FromStr for ColorChoice {
+    type Err = ColorChoiceParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if s == "always" {
+            Ok(ColorChoice::Always)
+        } else if s == "auto" {
+            Ok(ColorChoice::Auto)
+        } else if s == "never" {
+            Ok(ColorChoice::Never)
+        } else {
+            Err(ColorChoiceParseError)
+        }
+    }
+}
+
+#[derive(Debug, derive_more::Display)]
+#[display(fmt = "Color must be one of: always, auto, never")]
+pub struct ColorChoiceParseError;
